+++ docs/CHANGELOG.md
@@ -13,6 +13,10 @@ The versioning system is basically the following:
   - If there's a really "big" change.
 - As mentioned, the "major" version is 0.
 
+## main
+
+- Use multi-line types less often in error messages.
+
 ## v0.14.0
 
 - Warn when a `.sig` file doesn't contain 1 `signature`.